+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - RPC v0.4 `starknet_getTransactionReceipt` incorrect execution and finality status names
+- `pathfinder_getTransactionStatus` fails to parse v0.12.1 gateway replies
 
 ## [0.7.0] - 2023-07-27